refactor(service): register agent kinds through one AgentKindDefinition
The writing service was 960 lines of which most was generic: session rows,
durable runs, prompt/attach/stream, abort, approvals, compaction and rewind.
That part is now createAgentKindService in apps/service/src/agents/service.ts,
implemented once over an AgentKindDefinition (agents/kind.ts). A kind supplies
only what differs — minTier, defaults, replay policy, model allowlist,
capabilities, its 1:1 state row, runTurn and maintenance — and the writing
kind is agents/writing.ts.

agents/registry.ts is the one place a kind is registered: it derives the
request-context agentKinds map and is what the turn step resolves
agent.session.kind against, replacing the hand-written lazy delegate and the
separate AGENT_TURN_HANDLERS map. Entries restate minTier for the guards and
load the definition with a dynamic import, so the boot path is unchanged.
